Updated gradle, deps and libs, updated DI, fixed crash when downloading image and wallpaper setting, fixed bug with hiding content with photo info

+++ app/src/main/java/com/greylabsdev/pexwalls/data/dto/PhotoDto.kt
@@ -19,7 +19,7 @@ data class PhotoDto(
     val photographer: String,
     @ColumnInfo
     @SerializedName("photographer_id")
-    val photographerId: Int,
+    val photographerId: Long,
     @ColumnInfo
     @SerializedName("photographer_url")
     val photographerUrl: String,
